refactor(edit): move manageSectionHeader to more appropriate folder
