+++ app/ui/src/app/edit/components/shared/ManageSectionHeader.tsx
